fixup: rename component to match convention
